requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.1
asn1crypto==1.5.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coverage==6.4.2
cryptography==38.0.3
cssselect2==0.7.0
dj-database-url==0.5.0
Django==3.2
django-resized==1.0.2
django-simple-history==3.2.0
django-smart-selects==1.5.9
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
lxml==4.9.1
mysql-connector-python==8.0.29
oscrypto==1.3.0
pandas==1.5.2
Pillow==9.1.1
protobuf==3.20.1
psycopg2-binary==2.9.3
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyMySQL==1.0.2
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.64.1
tzdata==2022.6
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
whitenoise==6.1.0
xhtml2pdf==0.2.8
